+++ src/Ext/WeaponType/Hook.EBolt.cpp
@@ -3,11 +3,8 @@
 #include <Utilities/Macro.h>
 #include <Helpers/Macro.h>
 
-namespace BoltTemp
-{
-	PhobosMap<EBolt*, const WeaponTypeExt::ExtData*> boltWeaponTypeExt;
-	const WeaponTypeExt::ExtData* pType = nullptr;
-}
+PhobosMap<EBolt*, const WeaponTypeExt::ExtData*> WeaponTypeExt::BoltWeaponMap;
+const WeaponTypeExt::ExtData* WeaponTypeExt::BoltWeaponType = nullptr;
 
 DEFINE_HOOK(0x6FD494, TechnoClass_FireEBolt_SetExtMap_AfterAres, 0x7)
 {
@@ -17,7 +14,7 @@ DEFINE_HOOK(0x6FD494, TechnoClass_FireEBolt_SetExtMap_AfterAres, 0x7)
 	if (pWeapon)
 	{
 		auto const pWeaponExt = WeaponTypeExt::ExtMap.Find(pWeapon);
-		BoltTemp::boltWeaponTypeExt[pBolt] = pWeaponExt;
+		WeaponTypeExt::BoltWeaponMap[pBolt] = pWeaponExt;
 		pBolt->Lifetime = 1 << (Math::clamp(pWeaponExt->Bolt_Duration, 1, 31) - 1);
 	}
 
@@ -28,7 +25,7 @@ DEFINE_HOOK(0x4C2951, EBolt_DTOR, 0x5)
 {
 	GET(EBolt*, pBolt, ECX);
 
-	BoltTemp::boltWeaponTypeExt.erase(pBolt);
+	WeaponTypeExt::BoltWeaponMap.erase(pBolt);
 
 	return 0;
 }
@@ -38,11 +35,11 @@ DEFINE_HOOK(0x4C20BC, EBolt_DrawArcs, 0xB)
 	enum { DoLoop = 0x4C20C7, Break = 0x4C2400 };
 
 	GET_STACK(EBolt*, pBolt, 0x40);
-	BoltTemp::pType = BoltTemp::boltWeaponTypeExt.get_or_default(pBolt);
+	WeaponTypeExt::BoltWeaponType = WeaponTypeExt::BoltWeaponMap.get_or_default(pBolt);
 
 	GET_STACK(int, plotIndex, STACK_OFFSET(0x408, -0x3E0));
 
-	int arcCount = BoltTemp::pType ? BoltTemp::pType->Bolt_Arcs : 8;
+	int arcCount = WeaponTypeExt::BoltWeaponType ? WeaponTypeExt::BoltWeaponType->Bolt_Arcs : 8;
 
 	return plotIndex < arcCount ? DoLoop : Break;
 }
@@ -50,17 +47,17 @@ DEFINE_HOOK(0x4C20BC, EBolt_DrawArcs, 0xB)
 DEFINE_HOOK(0x4C24E4, Ebolt_DrawFist_Disable, 0x8)
 {
 	GET_STACK(EBolt*, pBolt, 0x40);
-	BoltTemp::pType = BoltTemp::boltWeaponTypeExt.get_or_default(pBolt);
+	WeaponTypeExt::BoltWeaponType = WeaponTypeExt::BoltWeaponMap.get_or_default(pBolt);
 
-	return (BoltTemp::pType && BoltTemp::pType->Bolt_Disable1) ? 0x4C2515 : 0;
+	return (WeaponTypeExt::BoltWeaponType && WeaponTypeExt::BoltWeaponType->Bolt_Disable1) ? 0x4C2515 : 0;
 }
 
 DEFINE_HOOK(0x4C25FD, Ebolt_DrawSecond_Disable, 0xA)
 {
-	return (BoltTemp::pType && BoltTemp::pType->Bolt_Disable2) ? 0x4C262A : 0;
+	return (WeaponTypeExt::BoltWeaponType && WeaponTypeExt::BoltWeaponType->Bolt_Disable2) ? 0x4C262A : 0;
 }
 
 DEFINE_HOOK(0x4C26EE, Ebolt_DrawThird_Disable, 0x8)
 {
-	return (BoltTemp::pType && BoltTemp::pType->Bolt_Disable3) ? 0x4C2710 : 0;
+	return (WeaponTypeExt::BoltWeaponType && WeaponTypeExt::BoltWeaponType->Bolt_Disable3) ? 0x4C2710 : 0;
 }