+++ docs/Whats-New.md
@@ -328,6 +328,10 @@ New:
 - Damage multiplier for different houses (by CrimRecya)
 - Customizable duration for electric bolts (by Starkku)
 - Extended gattling rate down logic (by CrimRecya)
+- Sell or undeploy building on impact (by CrimRecya)
+- No rearm and reload in EMP or temporal (by CrimRecya)
+- Draw visual effects for airburst weapons (by CrimRecya)
+- Technos recount current burst index when change the firing weapon (by CrimRecya)
 - New Engrave trajectory (by CrimRecya)
 
 Vanilla fixes: